move paragraph into components
